商品出品#4
Validationの修正
JSファイルの整理
